examples/mpu: support a test of application access and tidyup
1. Support APP_BINARY_SEPARATION to test application region protection
2. Remove unnecessary separated files, tc_read/write_code/data.c
3. Add prebuilt application binaries, micom and wifi for imxrt1050-evk
