+++ apps/examples/mpu/Kconfig
@@ -24,35 +24,25 @@ config MPUTEST_KERNEL_DATA_ADDR
 		Test MPU access protection to this kernel data address
 
 config MPUTEST_APP_ADDR
-	hex "Address in application range"
+	hex "Address in differnet application range"
 	default 0x00000000
+	depends on APP_BINARY_SEPARATION
 	---help---
 		Test MPU access protection to this application address
 
-
-config TC_MPU_ALL
-	bool "All"
+config MPUTEST_USE_PREBUILT_BINS
+	bool "Use pre-built application binaries"
 	default n
-	select TC_MPU_READ_CODE
-	select TC_MPU_WRITE_CODE
-	select TC_MPU_READ_DATA
-	select TC_MPU_WRITE_DATA
-
-config TC_MPU_READ_CODE
-	bool "Read Kernel Code"
-	default n
-
-config TC_MPU_WRITE_CODE
-	bool "Write Kernel Code"
-	default n
-
-config TC_MPU_READ_DATA
-	bool "Read Kernel Data"
-	default n
-
-config TC_MPU_WRITE_DATA
-	bool "Write Kernel Data"
-	default n
+	depends on APP_BINARY_SEPARATION
+	depends on ARCH_BOARD_IMXRT1050_EVK
+	depends on !EXAMPLES_ELF
+	---help---
+		To test different application access, two application binaries
+		should be flashed. This config makes using pre-built application binaries.
+		Note! 1. Initial binaries are built with NXP imxrt1050-evk. To use this
+		        with another board, pre-built binaries should be replaced.
+		      2. ELF example makes application binaries itself so that
+		         when it is enabled, this should be turned off.
 
 endif #EXAMPLES_MPU_TEST
